+++ service/v1/users/user_service.go
@@ -45,7 +45,7 @@ func (c *Client) GetUsage(userID string) (*model.UserUsage, error) {
 	usage := &model.UserUsage{
 		CpuCores: float64(vmUsage.CpuCores) + deploymentUsage.CpuCores,
 		RAM:      float64(vmUsage.RAM) + deploymentUsage.RAM,
-		DiskSize: float64(vmUsage.DiskSize),
+		DiskSize: vmUsage.DiskSize,
 	}
 
 	return usage, nil
